refactor: simplify concrete/api dir
